// src/services/generated/models/GithubRepoResponse.ts
export interface GithubRepoResponse {
    /**
     * 
     * @type {string}
     * @memberof GithubRepoResponse
     */
    description?: string;
    /**
     * 
     * @type {number}
     * @memberof GithubRepoResponse
     */
    forks?: number;
    /**
     * 
     * @type {string}
     * @memberof GithubRepoResponse
     */
    fullName?: string;
    /**
     * 
     * @type {string}
     * @memberof GithubRepoResponse
     */
    homepage?: string;
    /**
     * 
     * @type {string}
     * @memberof GithubRepoResponse
     */
    htmlUrl?: string;
    /**
     * 
     * @type {number}
     * @memberof GithubRepoResponse
     */
    id?: number;
    /**
     * 
     * @type {boolean}
     * @memberof GithubRepoResponse
     */
    isArchived?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof GithubRepoResponse
     */
    isFork?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof GithubRepoResponse
     */
    isPinned?: boolean;
    /**
     * 
     * @type {string}
     * @memberof GithubRepoResponse
     */
    language?: string;
    /**
     * 
     * @type {string}
     * @memberof GithubRepoResponse
     */
    name?: string;
    /**
     * 
     * @type {number}
     * @memberof GithubRepoResponse
     */
    openIssues?: number;
    /**
     * 
     * @type {string}
     * @memberof GithubRepoResponse
     */
    pushedAt?: string;
    /**
     * 
     * @type {number}
     * @memberof GithubRepoResponse
     */
    stars?: number;
    /**
     * 
     * @type {Array<string>}
     * @memberof GithubRepoResponse
     */
    topics?: Array<string>;
    /**
     * 
     * @type {number}
     * @memberof GithubRepoResponse
     */
    watchers?: number;
}